requirements.txt
numpy
scipy
pandas
psychopy
pyserial
matplotlib
seaborn
pingouin
papermill
systole
setuptools
packaging
cbitstruct
pylsl